+++ packages/hpl-nectar-missions/types.ts
@@ -1,0 +1,12 @@
+import { EarnedReward } from "./generated/types";
+
+export interface OffchainParticipation {
+  _id: string;
+  wallet: string;
+  mission: string;
+  nft: string;
+  endTime: string;
+  isRecalled: boolean;
+  rewards: EarnedReward[];
+  updatedAt: string;
+}