+++ packages/hpl-nectar-missions/utils/index.ts
@@ -1,3 +1,7 @@
+import { PublicKey } from "@solana/web3.js";
+import { Participation } from "../generated";
+import { OffchainParticipation } from "../types";
+
 export * from "./pdas";
 export * from "./merkleTree";
 export * from "./lookupTables";
@@ -18,3 +22,24 @@ export function removeDuplicateFromArrayOf<T = any, C = any>(
       })
   );
 }
+
+export const offchainToSolitaParticipation = (
+  participation: OffchainParticipation
+): Participation =>
+  Participation.fromArgs({
+    bump: 0,
+    wallet: new PublicKey(participation.wallet),
+    mission: new PublicKey(participation.mission),
+    nft: new PublicKey(participation.nft),
+    endTime: new Date(participation.endTime).getTime() / 1000,
+    isRecalled: participation.isRecalled,
+    rewards: participation.rewards.map((reward) => ({
+      ...reward,
+      rewardType: {
+        ...reward.rewardType,
+        address:
+          reward.rewardType.__kind == "Currency" &&
+          new PublicKey(reward.rewardType.address),
+      },
+    })),
+  });