+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.1.46
 
 - adds init_if_needed support for staker in stake operation
+- Adds Nft order in `NFTv1` Used by State for Use in Guilds
+
 
 ## 0.1.45